updating search for paper and discussions
